fix: rename custom button
